F1 light used for reference for subsequent task
This is the f1 light fully implemented. There are the following bugs. 1.) When uploaded to vbuddy and code running for the first time, you need to press three times until the lights start counting. 2.) When change the parameter caliN to 22 (or higher numbers) the code no longer runs. Having low caliN means that the counting is faster than 1 second specified by the Lab3 Task4 description.

Let me know if there are any other errors.